refactor: move tests from `polynomial`
Also implemented a `Display` for Lagrange basis polynomials.
